+++ composer.json
@@ -22,7 +22,8 @@
     "require": {
         "php": ">=7.1.0",
         "silverstripe/framework": "^4",
-        "silverstripe/vendor-plugin": "^1"
+        "silverstripe/vendor-plugin": "^1",
+        "silverstripe/event-dispatcher": "dev-master"
     },
     "require-dev": {
         "phpunit/phpunit": "^5.7",
