--- src/Dispatch/EventHandlerLoader.php
@@ -1,9 +1,0 @@
-<?php
-
-
-namespace SilverStripe\Snapshots\Dispatch;
-
-interface EventHandlerLoader
-{
-    public function addToDispatcher(Dispatcher $dispatcher): void;
-}
