+++ src/Handler/Form/PublishHandler.php
@@ -3,13 +3,13 @@
 
 namespace SilverStripe\Snapshots\Handler\Form;
 
+use SilverStripe\EventDispatcher\Event\EventContextInterface;
 use SilverStripe\Forms\Form;
-use SilverStripe\Snapshots\Listener\EventContext;
 use SilverStripe\Snapshots\Snapshot;
 
 class PublishHandler extends Handler
 {
-    protected function createSnapshot(EventContext $context): ?Snapshot
+    protected function createSnapshot(EventContextInterface $context): ?Snapshot
     {
         $snapshot = parent::createSnapshot($context);
         if (!$snapshot) {
