+++ src/Handler/GridField/Alteration/Handler.php
@@ -3,20 +3,20 @@
 
 namespace SilverStripe\Snapshots\Handler\GridField\Alteration;
 
+use SilverStripe\EventDispatcher\Event\EventContextInterface;
 use SilverStripe\Forms\Form;
 use SilverStripe\ORM\ValidationException;
 use SilverStripe\Snapshots\Handler\HandlerAbstract;
-use SilverStripe\Snapshots\Listener\EventContext;
 use SilverStripe\Snapshots\Snapshot;
 
 class Handler extends HandlerAbstract
 {
     /**
-     * @param EventContext $context
+     * @param EventContextInterface $context
      * @return Snapshot|null
      * @throws ValidationException
      */
-    protected function createSnapshot(EventContext $context): ?Snapshot
+    protected function createSnapshot(EventContextInterface $context): ?Snapshot
     {
         $action = $context->getAction();
         if ($action === null) {
