--- src/Handler/HandlerInterface.php
@@ -1,11 +1,0 @@
-<?php
-
-
-namespace SilverStripe\Snapshots\Handler;
-
-use SilverStripe\Snapshots\Listener\EventContext;
-
-interface HandlerInterface
-{
-    public function fire(EventContext $context): void;
-}
